+++ CHANGELOG.md
@@ -11,6 +11,7 @@ project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 ### Fixed
 * Fix dangling env pointer in image MIME data cleanup (#2550)
+* Set canvas size on addPage and don't destroy PDF surface on width/height change (#2538)
 
 3.2.1
 ==================